Compute the VACS Index 2.

Compute the VACS Index 2.

Usage

score_vacsindex2(age, cd4_count, vl_log, hgb, fib4, egfr, hepc, alb, bmi, wbc)

points_vacindex2()

Arguments

age

Numeric. Year.

cd4_count

Numeric. CD4 cell count. cell/mm\(^3\).

vl_log

Numeric. log 10 of viral load.

hgb

Numeric. Hemoglobin. g/dL.

fib4

Numeric. Fibrosis-4 Index for Liver Fibrosis.

egfr

Numeric. Estimated glomerular filtration rate. (ml/min/1.73m\(^2\)).

hepc

Logical. TRUE for co-infection with hepatitis C virus.

alb

Numeric. Albumin. g/dL.

bmi

Numeric. Body mass index. m/kg\(^2\).

wbc

Numeric. White blood cell. \(10^3\) count/µL.

Value

A numerical vector.

References

Tate, J. P., Sterne, J. A., Justice, A. C., Study, V. A. C., & Antiretroviral Therapy Cohort Collaboration (ART-CC. (2019). Albumin, white blood cell count, and body mass index improve discrimination of mortality in HIV-positive individuals. AIDS (London, England), 33(5), 903.

McGinnis, K. A., Justice, A. C., Moore, R. D., Silverberg, M. J., Althoff, K. N., Karris, M., ... & Study, V. A. C. (2021). Discrimination And Calibration Of The Vacs Index 2.0 For Predicting Mortality Among People With Hiv In North America. Clinical Infectious Diseases: an Official Publication of the Infectious Diseases Society of America.

Examples

data("vacs2")
vacs2$score_vacs2 <- score_vacsindex2(
  age = vacs2$AGE,cd4_count = vacs2$CD4,vl_log = vacs2$VL_LOG,
  hgb = vacs2$HGB,fib4 = vacs2$FIB4,egfr = vacs2$EGFR,
  hepc = vacs2$HEPC==1,alb = vacs2$ALB,bmi = vacs2$BMI,wbc = vacs2$WBC)